Make history op object field types consistent #2303

+++ libraries/plugins/apis/account_history_api/include/steem/plugins/account_history_api/account_history_api.hpp
@@ -33,8 +33,8 @@ struct api_operation_object
    steem::protocol::transaction_id_type trx_id;
    uint32_t                               block = 0;
    uint32_t                               trx_in_block = 0;
-   uint16_t                               op_in_trx = 0;
-   uint16_t                               virtual_op = 0;
+   uint32_t                               op_in_trx = 0;
+   uint32_t                               virtual_op = 0;
    fc::time_point_sec                     timestamp;
    steem::protocol::operation             op;
